refactor: Position 디렉토리 위치 이동 (조직->공고) #66
