+++ src/main/java/KUSITMS/WITHUS/domain/evaluation/evaluationCriteria/entity/EvaluationCriteria.java
@@ -1,6 +1,6 @@
 package KUSITMS.WITHUS.domain.evaluation.evaluationCriteria.entity;
 
-import KUSITMS.WITHUS.domain.application.position.entity.Position;
+import KUSITMS.WITHUS.domain.recruitment.position.entity.Position;
 import KUSITMS.WITHUS.domain.evaluation.evaluation.entity.Evaluation;
 import KUSITMS.WITHUS.domain.evaluation.evaluationCriteria.enumerate.EvaluationType;
 import KUSITMS.WITHUS.domain.recruitment.recruitment.entity.Recruitment;
