+++ src/main/java/KUSITMS/WITHUS/domain/interview/timeslot/entity/TimeSlot.java
@@ -1,7 +1,7 @@
 package KUSITMS.WITHUS.domain.interview.timeslot.entity;
 
 import KUSITMS.WITHUS.domain.application.application.entity.Application;
-import KUSITMS.WITHUS.domain.application.position.entity.Position;
+import KUSITMS.WITHUS.domain.recruitment.position.entity.Position;
 import KUSITMS.WITHUS.domain.interview.interview.entity.Interview;
 import KUSITMS.WITHUS.domain.interview.timeslotUser.entity.TimeSlotUser;
 import KUSITMS.WITHUS.global.common.BaseEntity;
