+++ src/main/java/KUSITMS/WITHUS/domain/recruitment/position/repository/PositionJpaRepository.java
@@ -1,6 +1,6 @@
-package KUSITMS.WITHUS.domain.application.position.repository;
+package KUSITMS.WITHUS.domain.recruitment.position.repository;
 
-import KUSITMS.WITHUS.domain.application.position.entity.Position;
+import KUSITMS.WITHUS.domain.recruitment.position.entity.Position;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PositionJpaRepository extends JpaRepository<Position, Long> {
