+++ src/main/java/KUSITMS/WITHUS/domain/recruitment/position/repository/PositionRepository.java
@@ -1,6 +1,6 @@
-package KUSITMS.WITHUS.domain.application.position.repository;
+package KUSITMS.WITHUS.domain.recruitment.position.repository;
 
-import KUSITMS.WITHUS.domain.application.position.entity.Position;
+import KUSITMS.WITHUS.domain.recruitment.position.entity.Position;
 
 import java.util.Optional;
