+++ src/main/java/KUSITMS/WITHUS/domain/recruitment/recruitment/entity/Recruitment.java
@@ -1,6 +1,6 @@
 package KUSITMS.WITHUS.domain.recruitment.recruitment.entity;
 
-import KUSITMS.WITHUS.domain.application.position.entity.Position;
+import KUSITMS.WITHUS.domain.recruitment.position.entity.Position;
 import KUSITMS.WITHUS.domain.evaluation.evaluationCriteria.enumerate.EvaluationScaleType;
 import KUSITMS.WITHUS.domain.recruitment.availableTimeRange.entity.AvailableTimeRange;
 import KUSITMS.WITHUS.domain.recruitment.documentQuestion.entity.DocumentQuestion;
