+++ src/main/java/KUSITMS/WITHUS/domain/recruitment/recruitment/service/RecruitmentServiceImpl.java
@@ -1,6 +1,6 @@
 package KUSITMS.WITHUS.domain.recruitment.recruitment.service;
 
-import KUSITMS.WITHUS.domain.application.position.entity.Position;
+import KUSITMS.WITHUS.domain.recruitment.position.entity.Position;
 import KUSITMS.WITHUS.domain.organization.organization.entity.Organization;
 import KUSITMS.WITHUS.domain.organization.organization.repository.OrganizationRepository;
 import KUSITMS.WITHUS.domain.recruitment.availableTimeRange.entity.AvailableTimeRange;
